+++ packages/plugin-metrics/src/node/metric-output/plugin-metrics-time-count.ts
@@ -23,12 +23,12 @@ export class PluginMetricTimeCount implements MetricOutput {
     public header = '# HELP language_server_time_count Number of language server requests\n# TYPE language_server_time_count gauge\n';
 
     createMetricOutput(id: string, method: string, requestAnalytics: AnalyticsFromRequests): string {
-        if (requestAnalytics.succesfulResponses < 0) {
-            requestAnalytics.succesfulResponses = 0;
+        if (requestAnalytics.successfulResponses < 0) {
+            requestAnalytics.successfulResponses = 0;
         }
-        const successMetric = `language_server_time_count{id="${id}" method="${method}" result="success"} ${requestAnalytics.succesfulResponses}\n`;
+        const successMetric = `language_server_time_count{id="${id}" method="${method}" result="success"} ${requestAnalytics.successfulResponses}\n`;
 
-        const failedRequests = requestAnalytics.totalRequests - requestAnalytics.succesfulResponses;
+        const failedRequests = requestAnalytics.totalRequests - requestAnalytics.successfulResponses;
         const failureMetric = `language_server_time_count{id="${id}" method="${method}" result="fail"} ${failedRequests}\n`;
         return successMetric + failureMetric;
     }